Refactor ViewModelFactories and token validation
This commit removes the `ChatViewModelFactory`, `LoginViewModelFactory`, `MainViewModelFactory`, and `StoryViewModelFactory` as they are no longer needed due to the integration of Koin for dependency injection.

It also introduces token validation in `LoginActivity` to check the validity of the stored token before navigating to `MainActivity`. If the token is invalid or expired, the user is prompted to log in again. The `MainViewModel` now accepts a `uId` parameter, which is passed from `LoginActivity` after successful token validation.

Other changes include:
- Added `TokenValidationResponse` data class.
- Updated `RetrofitInterfaceLogin` with a `validateToken` endpoint.
- Modified `LoginViewModel` to include a `validateToken` function and use `emit` for StateFlow updates.
- Updated various screens (Home, Search, ViewPost, UserProfile, MyProfile, Story, Settings, Chat, NewMessage, Messages) to reflect changes in `MainViewModel` instantiation.
- In `Home.kt`, simplified image loading for the user's profile picture and updated `fetchFeed` to not require `uId` as a parameter.
- `MainViewModel`'s `setPosts` method no longer requires `uId` as a parameter.